reformat tests using requestBuilder
