+++ explorer/src/__tests__/controllers/admin/nodes.test.ts
@@ -1,4 +1,3 @@
-import request from 'supertest'
 import http from 'http'
 import httpStatus from 'http-status-codes'
 import { Connection } from 'typeorm'
@@ -10,10 +9,7 @@ import {
   find as findNode,
 } from '../../../entity/ChainlinkNode'
 import { start, stop } from '../../../support/server'
-import {
-  ADMIN_USERNAME_HEADER,
-  ADMIN_PASSWORD_HEADER,
-} from '../../../utils/constants'
+import { requestBuilder, RequestBuilder } from '../../../support/requestBuilder'
 
 const USERNAME = 'myadmin'
 const PASSWORD = 'validpassword'
@@ -22,10 +18,12 @@ const adminNodesPath = `${ADMIN_PATH}/nodes`
 
 let server: http.Server
 let db: Connection
+let rb: RequestBuilder
 
 beforeAll(async () => {
   db = await getDb()
   server = await start()
+  rb = requestBuilder(server)
 })
 afterAll(done => stop(server, done))
 beforeEach(async () => {
@@ -33,43 +31,11 @@ beforeEach(async () => {
   await createAdmin(db, USERNAME, PASSWORD)
 })
 
-function sendGet(path: string, username: string, password: string) {
-  return request(server)
-    .get(path)
-    .set('Accept', 'application/json')
-    .set('Content-Type', 'application/json')
-    .set(ADMIN_USERNAME_HEADER, username)
-    .set(ADMIN_PASSWORD_HEADER, password)
-}
-
-function sendPost(
-  path: string,
-  data: object,
-  username: string,
-  password: string,
-) {
-  return request(server)
-    .post(path)
-    .send(data)
-    .set('Accept', 'application/json')
-    .set('Content-Type', 'application/json')
-    .set(ADMIN_USERNAME_HEADER, username)
-    .set(ADMIN_PASSWORD_HEADER, password)
-}
-
-function sendDelete(path: string, username: string, password: string) {
-  return request(server)
-    .delete(path)
-    .set('Content-Type', 'application/json')
-    .set(ADMIN_USERNAME_HEADER, username)
-    .set(ADMIN_PASSWORD_HEADER, password)
-}
-
 describe('POST /api/v1/admin/nodes', () => {
   it('can create a node and returns the generated information', done => {
     const data = { name: 'nodeA', url: 'http://nodea.com' }
 
-    sendPost(adminNodesPath, data, USERNAME, PASSWORD)
+    rb.sendPost(adminNodesPath, USERNAME, PASSWORD, data)
       .expect(httpStatus.CREATED)
       .expect(res => {
         expect(res.body.id).toBeDefined()
@@ -82,7 +48,7 @@ describe('POST /api/v1/admin/nodes', () => {
   it('returns an error with invalid params', done => {
     const data = { url: 'http://nodea.com' }
 
-    sendPost(adminNodesPath, data, USERNAME, PASSWORD)
+    rb.sendPost(adminNodesPath, USERNAME, PASSWORD, data)
       .expect(httpStatus.UNPROCESSABLE_ENTITY)
       .expect(res => {
         const errors = res.body.errors
@@ -99,13 +65,13 @@ describe('POST /api/v1/admin/nodes', () => {
     const [node] = await createChainlinkNode(db, 'nodeA')
     const data = { name: node.name }
 
-    sendPost(adminNodesPath, data, USERNAME, PASSWORD)
+    rb.sendPost(adminNodesPath, USERNAME, PASSWORD, data)
       .expect(httpStatus.CONFLICT)
       .end(done)
   })
 
   it('returns a 401 unauthorized with invalid admin credentials', done => {
-    sendPost(adminNodesPath, {}, USERNAME, 'invalidpassword')
+    rb.sendPost(adminNodesPath, USERNAME, 'invalidpassword')
       .expect(httpStatus.UNAUTHORIZED)
       .end(done)
   })
@@ -119,7 +85,7 @@ describe('DELETE /api/v1/admin/nodes/:name', () => {
   it('can delete a node', async done => {
     const [node] = await createChainlinkNode(db, 'nodeA')
 
-    sendDelete(path(node.name), USERNAME, PASSWORD)
+    rb.sendDelete(path(node.name), USERNAME, PASSWORD)
       .expect(httpStatus.OK)
       .expect(async () => {
         const nodeAfter = await findNode(db, node.id)
@@ -129,7 +95,7 @@ describe('DELETE /api/v1/admin/nodes/:name', () => {
   })
 
   it('returns a 401 unauthorized with invalid admin credentials', done => {
-    sendDelete(path('idontexist'), USERNAME, 'invalidpassword')
+    rb.sendDelete(path('idontexist'), USERNAME, 'invalidpassword')
       .expect(httpStatus.UNAUTHORIZED)
       .end(done)
   })
@@ -143,7 +109,7 @@ describe.only('GET /api/v1/admin/nodes/:id', () => {
   it('can get a node', async done => {
     const [node] = await createChainlinkNode(db, 'nodeA')
 
-    sendGet(path(node.id), USERNAME, PASSWORD)
+    rb.sendGet(path(node.id), USERNAME, PASSWORD)
       .expect(httpStatus.OK)
       .expect(res => {
         console.log(res.body)
@@ -155,7 +121,7 @@ describe.only('GET /api/v1/admin/nodes/:id', () => {
   it('returns a 401 unauthorized with invalid admin credentials', async done => {
     const [node] = await createChainlinkNode(db, 'nodeA')
     const _nodePath = path(node.id)
-    sendGet(_nodePath, USERNAME, 'invalidpassword')
+    rb.sendGet(_nodePath, USERNAME, 'invalidpassword')
       .expect(httpStatus.UNAUTHORIZED)
       .end(done)
   })
